+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - implemented better support for themes + new themes added
 - implemented support for fonts + new fonts added
 - added keyboard shortcuts to toggle TODO list (Alt+T), Include context files (Alt+F), Include repository map (Alt+M), and Auto-approve (Alt+Y)
+- updated application main menu and included Settings
 
 ## [0.26.0]